requirements.txt
# Django
Django>=5.0,<6.0
django-environ>=0.11.2

# GraphQL
graphene-django>=3.2.0
PyJWT>=2.8.0

# Database
psycopg2-binary>=2.9.9

# Twilio
twilio>=9.0.0

# Utilities
phonenumbers>=8.13.0
